requirements.txt
dash==2.14.2
pandas==2.2.3
plotly==5.18.0
kaleido==0.2.1
python-docx==1.0.1
reportlab==4.0.8